+++ src/idc/reader/__init__.py
@@ -1,3 +1,4 @@
 from ._data import DataReader
+from ._multi import MultiReader
 from ._poll_dir import PollDir
 from ._pyfunc import PythonFunctionReader