+++ src/idc/writer/__init__.py
@@ -1,2 +1,3 @@
 from ._data import DataWriter
+from ._multi import MultiWriter
 from ._pyfunc import PythonFunctionWriter